Rename fritz coordinator module
