+++ homeassistant/components/fritz/button.py
@@ -19,8 +19,14 @@ from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC, Device
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .common import AvmWrapper, FritzData, FritzDevice, FritzDeviceBase, _is_tracked
 from .const import BUTTON_TYPE_WOL, CONNECTION_TYPE_LAN, DATA_FRITZ, DOMAIN, MeshRoles
+from .coordinator import (
+    AvmWrapper,
+    FritzData,
+    FritzDevice,
+    FritzDeviceBase,
+    _is_tracked,
+)
 
 _LOGGER = logging.getLogger(__name__)
