+++ homeassistant/components/fritz/switch.py
@@ -17,15 +17,6 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from homeassistant.util import slugify
 
-from .common import (
-    AvmWrapper,
-    FritzBoxBaseEntity,
-    FritzData,
-    FritzDevice,
-    FritzDeviceBase,
-    SwitchInfo,
-    device_filter_out_from_trackers,
-)
 from .const import (
     DATA_FRITZ,
     DOMAIN,
@@ -35,6 +26,15 @@ from .const import (
     SWITCH_TYPE_WIFINETWORK,
     WIFI_STANDARD,
     MeshRoles,
+)
+from .coordinator import (
+    AvmWrapper,
+    FritzBoxBaseEntity,
+    FritzData,
+    FritzDevice,
+    FritzDeviceBase,
+    SwitchInfo,
+    device_filter_out_from_trackers,
 )
 
 _LOGGER = logging.getLogger(__name__)
