+++ homeassistant/components/fritz/update.py
@@ -16,8 +16,12 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .common import AvmWrapper, FritzBoxBaseCoordinatorEntity, FritzEntityDescription
 from .const import DOMAIN
+from .coordinator import (
+    AvmWrapper,
+    FritzBoxBaseCoordinatorEntity,
+    FritzEntityDescription,
+)
 
 _LOGGER = logging.getLogger(__name__)
