+++ tests/components/fritz/conftest.py
@@ -84,7 +84,7 @@ def fc_data_mock():
 def fc_class_mock(fc_data):
     """Fixture that sets up a mocked FritzConnection class."""
     with patch(
-        "homeassistant.components.fritz.common.FritzConnection", autospec=True
+        "homeassistant.components.fritz.coordinator.FritzConnection", autospec=True
     ) as result:
         result.return_value = FritzConnectionMock(fc_data)
         yield result
@@ -94,7 +94,7 @@ def fc_class_mock(fc_data):
 def fh_class_mock():
     """Fixture that sets up a mocked FritzHosts class."""
     with patch(
-        "homeassistant.components.fritz.common.FritzHosts",
+        "homeassistant.components.fritz.coordinator.FritzHosts",
         new=FritzHosts,
     ) as result:
         result.get_mesh_topology = MagicMock(return_value=MOCK_MESH_DATA)
