+++ tests/components/fritz/test_init.py
@@ -76,7 +76,7 @@ async def test_setup_auth_fail(hass: HomeAssistant, error) -> None:
     entry.add_to_hass(hass)
 
     with patch(
-        "homeassistant.components.fritz.common.FritzConnection",
+        "homeassistant.components.fritz.coordinator.FritzConnection",
         side_effect=error,
     ):
         await hass.config_entries.async_setup(entry.entry_id)
@@ -96,7 +96,7 @@ async def test_setup_fail(hass: HomeAssistant, error) -> None:
     entry.add_to_hass(hass)
 
     with patch(
-        "homeassistant.components.fritz.common.FritzConnection",
+        "homeassistant.components.fritz.coordinator.FritzConnection",
         side_effect=error,
     ):
         await hass.config_entries.async_setup(entry.entry_id)
